Remove dependency on goquery and clear go.mod requires
This is a rewrite of the upgrade/upgrade.go script (which is used only
by maintainers) to use only packages from stdlib, and in particular
remove the use of the external dependency on
github.com/PuerkitoBio/goquery.

As a result, the go.mod is now cleaned from any external dependency and
all Go modules that depend on go-sqlite3 will have the goquery
dependency cleaned from their dependency tree (go.sum).