sqlite3: don't copy string in bind()
Change bind() to pass sqlite3_bind_text() a pointer to the strings data
instead of converting it to a []byte just so a pointer to `&b[0]` can be
passed to unsafe.Pointer. Basically, this saves a needless allocation.
and passing a pointer to that.

This is safe because sqlite3_bind_text does not keep a copy of the
provided string.