bugfix: after executing statement table lock was not cleared
When we execute a statement, we step once and return. If the query
returned more than one row, we would not have consumed everything so the
underlying table would still be locked even when the execution had
finished.

This change resets the statement after execution so that the locks can
be cleared.