sqlite3_opt_preupdate_hook: ensure backward compat with Segment fork
Prior to February 2024, this fork contained several changes:

- Fix to a race condition; this was added in 2018 and merged upstream.
- Add a pre-update hook; this was merged _behind a build tag_
- Various changes to update the build infrastructure or the sqlite version.

As of February 2024, _the only change from upstream removes the
build tag_ from the pre-update hook. This is to preserve backward
compatibility, and to allow us to update our Segment internal code
without needing to add a Go build tag to every project that includes
this library.

This commit replaces the fork's contents before February 2024. This
version should be easier to maintain going forward.